Add files via upload
The Temperature Converter is a common tool used in daily life and has various applications in different fields such as meteorology, thermodynamics, and physics. With the rise of computer technology, it is now possible to build a graphical user interface (GUI) for this tool, making it easier and more convenient to use. In this project, we will be building a Temperature Converter in Java using the Swing library. Swing is a rich set of graphical user interface components that can be used to create a variety of graphical applications. Here we created a temperature converter with the Swing library in the Java programming language.